+++ spec/fixtures/response/response_fixture.rb
@@ -1,36 +1,36 @@
 class ResponseFixture
+  class << self
 
-  def self.authentication(value = nil)
-    case value
-      when :to_hash
-        { :success => true,
-          :authentication_value => {
-            :token => "a68d1d6379b62ff339a0e0c69ed4d9cf",
-            :token_hash => "AAAJxA;cIedoT;mY10ExZwG6JuKgp2OYKxow==",
-            :client => "radclient"
-          }
-        }
-      else
-        @@authentication ||= load_fixture :authentication
+    def authentication(value = nil)
+      @authentication ||= load_fixture :authentication
+      
+      case value
+        when :to_hash then Savon::SOAP::XML.to_hash(@authentication)[:authenticate_response][:return]
+        else               @authentication
+      end
     end
+
+    def soap_fault
+      @soap_fault ||= load_fixture :soap_fault
+    end
+
+    def soap_fault12
+      @soap_fault12 ||= load_fixture :soap_fault12
+    end
+
+    def multi_ref
+      @multi_ref ||= load_fixture :multi_ref
+    end
+
+    def list
+      @list ||= load_fixture :list
+    end
+
+  private
+
+    def load_fixture(fixture)
+      File.read File.dirname(__FILE__) + "/xml/#{fixture}.xml"
+    end
+
   end
-
-  def self.soap_fault
-    @@soap_fault ||= load_fixture :soap_fault
-  end
-
-  def self.soap_fault12
-    @@soap_fault12 ||= load_fixture :soap_fault12
-  end
-
-  def self.multi_ref
-    @@multi_ref ||= load_fixture :multi_ref
-  end
-
-private
-
-  def self.load_fixture(fixture)
-    File.read File.dirname(__FILE__) + "/xml/#{fixture}.xml"
-  end
-
 end